Restructure user queries
