+++ pkg/registry/apis/identity/legacy/user.go
@@ -1,0 +1,111 @@
+package legacy
+
+import (
+	"context"
+	"fmt"
+	"text/template"
+
+	"github.com/grafana/authlib/claims"
+	"github.com/grafana/grafana/pkg/registry/apis/identity/common"
+	"github.com/grafana/grafana/pkg/services/user"
+	"github.com/grafana/grafana/pkg/storage/legacysql"
+	"github.com/grafana/grafana/pkg/storage/unified/sql/sqltemplate"
+)
+
+type ListUserQuery struct {
+	OrgID            int64
+	UID              string
+	IsServiceAccount bool
+
+	Pagination common.Pagination
+}
+
+type ListUserResult struct {
+	Users    []user.User
+	Continue int64
+	RV       int64
+}
+
+var sqlQueryUsersTemplate = mustTemplate("users_query.sql")
+
+func newListUser(sql *legacysql.LegacyDatabaseHelper, q *ListUserQuery) listUsersQuery {
+	return listUsersQuery{
+		SQLTemplate:  sqltemplate.New(sql.DialectForDriver()),
+		UserTable:    sql.Table("user"),
+		OrgUserTable: sql.Table("org_user"),
+		Query:        q,
+	}
+}
+
+type listUsersQuery struct {
+	sqltemplate.SQLTemplate
+	Query        *ListUserQuery
+	UserTable    string
+	OrgUserTable string
+}
+
+func (r listUsersQuery) Validate() error {
+	return nil // TODO
+}
+
+// ListUsers implements LegacyIdentityStore.
+func (s *legacySQLStore) ListUsers(ctx context.Context, ns claims.NamespaceInfo, query ListUserQuery) (*ListUserResult, error) {
+	// for continue
+	limit := int(query.Pagination.Limit)
+	query.Pagination.Limit += 1
+
+	query.OrgID = ns.OrgID
+	if ns.OrgID == 0 {
+		return nil, fmt.Errorf("expected non zero orgID")
+	}
+
+	sql, err := s.sql(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	res, err := s.queryUsers(ctx, sql, sqlQueryUsersTemplate, newListUser(sql, &query), limit)
+	if err == nil && query.UID != "" {
+		res.RV, err = sql.GetResourceVersion(ctx, "user", "updated")
+	}
+
+	return res, err
+}
+
+func (s *legacySQLStore) queryUsers(ctx context.Context, sql *legacysql.LegacyDatabaseHelper, t *template.Template, req sqltemplate.Args, limit int) (*ListUserResult, error) {
+	q, err := sqltemplate.Execute(t, req)
+	if err != nil {
+		return nil, fmt.Errorf("execute template %q: %w", t.Name(), err)
+	}
+
+	res := &ListUserResult{}
+	rows, err := sql.DB.GetSqlxSession().Query(ctx, q, req.GetArgs()...)
+	defer func() {
+		if rows != nil {
+			_ = rows.Close()
+		}
+	}()
+
+	if err == nil {
+		var lastID int64
+		for rows.Next() {
+			u := user.User{}
+			err = rows.Scan(&u.OrgID, &u.ID, &u.UID, &u.Login, &u.Email, &u.Name,
+				&u.Created, &u.Updated, &u.IsServiceAccount, &u.IsDisabled, &u.IsAdmin,
+			)
+			if err != nil {
+				return res, err
+			}
+
+			lastID = u.ID
+			res.Users = append(res.Users, u)
+			if len(res.Users) > limit {
+				res.Users = res.Users[0 : len(res.Users)-1]
+				res.Continue = lastID
+				break
+			}
+		}
+	}
+
+	return res, err
+}
